requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
django-crispy-forms==1.13.0
django-knox-rest==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
Pillow==8.4.0
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
